Added and fixed tests, all passing;
Added conftest.py in Registry directory, sets up fixtures for registry test.
